moved examples directory
